[ML] Remove abstract interface for reverse search creation
In previous products we used to create reverse searches
for more than one search engine in our categorization
functionality.  Realistically we will only ever create
reverse searches for Elasticsearch in the future, so
having an abstract interface for this functionality is
an unnecessary complication.  This PR removes it.
